+++ tests/VideoProcessingService.UnitTests/Tools/TempFileFixture.cs
@@ -4,6 +4,8 @@
     {
         public string FilePath { get; private set; }
 
+        private bool _disposed;
+
         public TempFileFixture(string path)
         {
             FilePath = path;
@@ -12,10 +14,21 @@
 
         public void Dispose()
         {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+        
+        protected virtual void Dispose(bool disposing)
+        {
+            if (_disposed) return;
+            if (!disposing) return;
+            
             if (File.Exists(FilePath))
             {
                 File.Delete(FilePath);
             }
+
+            _disposed = true;
         }
     }
 }
